+++ test/e2e/generate/generate_test.go
@@ -3,180 +3,17 @@ package generate
 import (
 	"errors"
 	"fmt"
-	"os"
 	"testing"
 	"time"
 
 	"github.com/kyverno/kyverno/test/e2e"
 	commonE2E "github.com/kyverno/kyverno/test/e2e/common"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"sigs.k8s.io/yaml"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
-
-func setup(t *testing.T) {
-	RegisterTestingT(t)
-	if os.Getenv("E2E") == "" {
-		t.Skip("Skipping E2E Test")
-	}
-}
-
-func createClient() *e2e.E2EClient {
-	client, err := e2e.NewE2EClient()
-	Expect(err).NotTo(HaveOccurred())
-	return client
-}
-
-func deleteClusteredResource(client *e2e.E2EClient, resource expectedResource) {
-	By(fmt.Sprintf("Deleting %s : %s", resource.gvr.String(), resource.name))
-	client.DeleteClusteredResource(resource.gvr, resource.name)
-	err := e2e.GetWithRetry(1*time.Second, 15, func() error {
-		_, err := client.GetClusteredResource(resource.gvr, resource.name)
-		if err == nil {
-			return fmt.Errorf("resource still exists: %s, %s", resource.gvr.String(), resource.name)
-		}
-		if apierrors.IsNotFound(err) {
-			return nil
-		}
-		return err
-	})
-	Expect(err).NotTo(HaveOccurred())
-}
-
-func deleteNamespacedResource(client *e2e.E2EClient, resource expectedResource) {
-	By(fmt.Sprintf("Deleting %s : %s/%s", resource.gvr.String(), resource.ns, resource.name))
-	client.DeleteNamespacedResource(resource.gvr, resource.ns, resource.name)
-	err := e2e.GetWithRetry(1*time.Second, 15, func() error {
-		_, err := client.GetNamespacedResource(resource.gvr, resource.ns, resource.name)
-		if err == nil {
-			return fmt.Errorf("resource still exists: %s, %s/%s", resource.gvr.String(), resource.ns, resource.name)
-		}
-		if apierrors.IsNotFound(err) {
-			return nil
-		}
-		return err
-	})
-	Expect(err).NotTo(HaveOccurred())
-}
-
-func deleteResource(client *e2e.E2EClient, resource expectedResource) {
-	if resource.ns != "" {
-		deleteNamespacedResource(client, resource)
-	} else {
-		deleteClusteredResource(client, resource)
-	}
-}
-
-func deleteResources(client *e2e.E2EClient, resources ...expectedResource) {
-	for _, resource := range resources {
-		deleteResource(client, resource)
-	}
-}
-
-func createClusteredResource(t *testing.T, client *e2e.E2EClient, resource resource) *unstructured.Unstructured {
-	var u unstructured.Unstructured
-	Expect(yaml.Unmarshal(resource.raw, &u)).To(Succeed())
-	By(fmt.Sprintf("Creating %s : %s", resource.gvr.String(), u.GetName()))
-	result, err := client.CreateClusteredResource(resource.gvr, &u)
-	Expect(err).NotTo(HaveOccurred())
-	t.Cleanup(func() {
-		deleteResources(client, expectedResource{resource.gvr, result.GetNamespace(), result.GetName()})
-	})
-	return result
-}
-
-func createNamespacedResource(t *testing.T, client *e2e.E2EClient, resource resource) *unstructured.Unstructured {
-	var u unstructured.Unstructured
-	Expect(yaml.Unmarshal(resource.raw, &u)).To(Succeed())
-	By(fmt.Sprintf("Creating %s : %s/%s", resource.gvr.String(), resource.ns, u.GetName()))
-	result, err := client.CreateNamespacedResource(resource.gvr, resource.ns, &u)
-	Expect(err).NotTo(HaveOccurred())
-	t.Cleanup(func() {
-		deleteResources(client, expectedResource{resource.gvr, result.GetNamespace(), result.GetName()})
-	})
-	return result
-}
-
-func createResource(t *testing.T, client *e2e.E2EClient, resource resource) *unstructured.Unstructured {
-	if resource.ns != "" {
-		return createNamespacedResource(t, client, resource)
-	} else {
-		return createClusteredResource(t, client, resource)
-	}
-}
-
-func createResources(t *testing.T, client *e2e.E2EClient, resources ...resource) {
-	for _, resource := range resources {
-		createResource(t, client, resource)
-	}
-}
-
-func expectClusteredResource(client *e2e.E2EClient, resource expectedResource) {
-	By(fmt.Sprintf("Expecting %s : %s", resource.gvr.String(), resource.name))
-	err := e2e.GetWithRetry(1*time.Second, 30, func() error {
-		_, err := client.GetClusteredResource(resource.gvr, resource.name)
-		if err != nil {
-			return err
-		}
-		return nil
-	})
-	Expect(err).NotTo(HaveOccurred())
-}
-
-func expectNamespacedResource(client *e2e.E2EClient, resource expectedResource) {
-	By(fmt.Sprintf("Expecting %s : %s/%s", resource.gvr.String(), resource.ns, resource.name))
-	err := e2e.GetWithRetry(1*time.Second, 30, func() error {
-		_, err := client.GetNamespacedResource(resource.gvr, resource.ns, resource.name)
-		if err != nil {
-			return err
-		}
-		return nil
-	})
-	Expect(err).NotTo(HaveOccurred())
-}
-
-func expectResource(client *e2e.E2EClient, resource expectedResource) {
-	if resource.ns != "" {
-		expectNamespacedResource(client, resource)
-	} else {
-		expectClusteredResource(client, resource)
-	}
-}
-
-func expectResources(client *e2e.E2EClient, resources ...expectedResource) {
-	for _, resource := range resources {
-		expectResource(client, resource)
-	}
-}
-
-func expectClusteredResourceNotExists(client *e2e.E2EClient, resource expectedResource) {
-	By(fmt.Sprintf("Expecting not exists %s : %s", resource.gvr.String(), resource.name))
-	_, err := client.GetClusteredResource(resource.gvr, resource.name)
-	Expect(apierrors.IsNotFound(err)).To(BeTrue())
-}
-
-func expectNamespacedResourceNotExists(client *e2e.E2EClient, resource expectedResource) {
-	By(fmt.Sprintf("Expecting not exists %s : %s/%s", resource.gvr.String(), resource.ns, resource.name))
-	_, err := client.GetNamespacedResource(resource.gvr, resource.ns, resource.name)
-	Expect(apierrors.IsNotFound(err)).To(BeTrue())
-}
-
-func expectResourceNotExists(client *e2e.E2EClient, resource expectedResource) {
-	if resource.ns != "" {
-		expectNamespacedResourceNotExists(client, resource)
-	} else {
-		expectClusteredResourceNotExists(client, resource)
-	}
-}
-
-func expectResourcesNotExist(client *e2e.E2EClient, resources ...expectedResource) {
-	for _, resource := range resources {
-		expectResourceNotExists(client, resource)
-	}
-}
 
 func Test_ClusterRole_ClusterRoleBinding_Sets(t *testing.T) {
 	setup(t)
@@ -241,92 +78,30 @@ func Test_Role_RoleBinding_Sets(t *testing.T) {
 func Test_Generate_NetworkPolicy(t *testing.T) {
 	setup(t)
 
-	// Generate E2E Client ==================
-	e2eClient, err := e2e.NewE2EClient()
-	Expect(err).To(BeNil())
-	// ======================================
+	for _, tests := range NetworkPolicyGenerateTests {
+		t.Run(tests.TestName, func(t *testing.T) {
+			e2eClient := createClient()
 
-	// ====== Range Over RuleTest ==================
-	for _, test := range NetworkPolicyGenerateTests {
-		By(fmt.Sprintf("Test to generate NetworkPolicy : %s", test.TestName))
-		By(fmt.Sprintf("synchronize = %v\t clone = %v", test.Sync, test.Clone))
+			t.Cleanup(func() {
+				deleteResources(e2eClient, tests.ExpectedResources...)
+			})
 
-		// ======= CleanUp Resources =====
-		By("Cleaning Cluster Policies")
-		_ = e2eClient.CleanClusterPolicies(clPolGVR)
+			// sanity check
+			expectResourcesNotExist(e2eClient, tests.ExpectedResources...)
 
-		// Clear Namespace
-		By(fmt.Sprintf("Deleting Namespace : %s", test.ResourceNamespace))
-		_ = e2eClient.DeleteClusteredResource(nsGVR, test.ResourceNamespace)
+			// create policy
+			policy := createResource(t, e2eClient, tests.ClusterPolicy)
+			Expect(commonE2E.PolicyCreated(policy.GetName())).To(Succeed())
 
-		// Wait Till Deletion of Namespace
-		err = e2e.GetWithRetry(1*time.Second, 15, func() error {
-			_, err := e2eClient.GetClusteredResource(nsGVR, test.ResourceNamespace)
-			if err != nil {
-				return nil
-			}
-			return errors.New("deleting Namespace")
+			// create source resources
+			createResources(t, e2eClient, tests.SourceResources...)
+
+			// create trigger
+			createResource(t, e2eClient, tests.TriggerResource)
+
+			// verify expected resources
+			expectResources(e2eClient, tests.ExpectedResources...)
 		})
-		Expect(err).NotTo(HaveOccurred())
-
-		// ======== Create Generate NetworkPolicy Policy =============
-		By("Creating Generate NetworkPolicy Policy")
-		_, err := e2eClient.CreateNamespacedResourceYaml(clPolGVR, npPolNS, test.PolicyName, test.Data)
-		Expect(err).NotTo(HaveOccurred())
-
-		err = commonE2E.PolicyCreated(test.PolicyName)
-		Expect(err).NotTo(HaveOccurred())
-
-		// ======= Create Namespace ==================
-		By(fmt.Sprintf("Creating Namespace which triggers generate %s", npPolNS))
-		_, err = e2eClient.CreateClusteredResourceYaml(nsGVR, namespaceWithLabelYaml)
-		Expect(err).NotTo(HaveOccurred())
-
-		// Wait Till Creation of Namespace
-		err = e2e.GetWithRetry(1*time.Second, 15, func() error {
-			_, err := e2eClient.GetClusteredResource(nsGVR, test.ResourceNamespace)
-			if err != nil {
-				return err
-			}
-			return nil
-		})
-		Expect(err).NotTo(HaveOccurred())
-
-		// ======== NetworkPolicy Creation =====
-		By(fmt.Sprintf("Verifying NetworkPolicy in the Namespace : %s", test.ResourceNamespace))
-
-		// Wait Till Creation of NetworkPolicy
-		err = e2e.GetWithRetry(1*time.Second, 15, func() error {
-			_, err := e2eClient.GetNamespacedResource(npGVR, test.ResourceNamespace, test.NetworkPolicyName)
-			if err != nil {
-				return err
-			}
-			return nil
-		})
-		Expect(err).NotTo(HaveOccurred())
-
-		npRes, err := e2eClient.GetNamespacedResource(npGVR, test.ResourceNamespace, test.NetworkPolicyName)
-		Expect(err).NotTo(HaveOccurred())
-		Expect(npRes.GetName()).To(Equal(test.NetworkPolicyName))
-		// ============================================
-
-		// ======= CleanUp Resources =====
-		_ = e2eClient.CleanClusterPolicies(clPolGVR)
-
-		// Clear Namespace
-		_ = e2eClient.DeleteClusteredResource(nsGVR, test.ResourceNamespace)
-
-		// Wait Till Deletion of Namespace
-		err = e2e.GetWithRetry(1*time.Second, 15, func() error {
-			_, err := e2eClient.GetClusteredResource(nsGVR, test.ResourceNamespace)
-			if err != nil {
-				return nil
-			}
-			return errors.New("deleting Namespace")
-		})
-		Expect(err).NotTo(HaveOccurred())
-
-		By(fmt.Sprintf("Test %s Completed \n\n\n", test.TestName))
 	}
 }
